Extract project provider context into a single hook
Since we don't store any state in the provider and mainly use it to
expose the server state via `useQuery` it will be easier if we fully
omit the provider.
