+++ application/ui/src/routes/datasets/record/record.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react';
 
 import { SchemaProjectConfig } from '../../../api/openapi-spec';
-import { useProject } from '../../projects/project.provider';
+import { useProject } from '../../../features/projects/use-project';
 import { HardwareSetup } from './hardware-setup';
 
 export const Record = () => {
-    const { project: projectConfig } = useProject();
+    const projectConfig = useProject();
     const [project, setProject] = useState<SchemaProjectConfig>(projectConfig);
 
     return <HardwareSetup project={project} setProject={setProject} />;
